% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeOptima.R
\name{computeOptima}
\alias{computeOptima}
\title{Use matrix bigraph to find matrix optimum row and column subsets.}
\usage{
computeOptima(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\description{
This function applies \code{\link{computeBicliques}} the bigraph constructed
from the row and column indices of a given matrix, wherein a row and column
are tied if their intersection is a non-missing cell.
}

