% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findOptimal.R
\name{findOptima}
\alias{findOptima}
\title{Seek optimum row and column subsets}
\usage{
findOptima(mat, pareto = FALSE, to.remove = FALSE, inf.na = TRUE)
}
\arguments{
\item{mat}{A matrix.}

\item{pareto}{Logical; whether to restrict the results to Pareto-optimal
subsets.}

\item{to.remove}{Logical; whether to return subsets of rows and columns to
keep (\code{FALSE}) or to remove (\code{TRUE}).}

\item{inf.na}{Logical; whether to treat infinite entries as undefined.
Defaults to \code{TRUE}.}
}
\description{
This function applies the heuristic \code{\link{ladderOptima}} search to both
a given matrix and its transpose, and synthesizes the results into optimal
row and colmun subsets.
}

