% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ladderOptima.R
\name{ladderOptima}
\alias{ladderOptima}
\title{Seek optimum row subsets}
\usage{
ladderOptima(mat, pareto = FALSE, to.remove = FALSE)
}
\arguments{
\item{mat}{A matrix.}

\item{pareto}{Logical; whether to restrict the results to Pareto-optimal
subsets.}

\item{to.remove}{Logical; whether to return subsets of rows and columns to
keep (\code{FALSE}) or to remove (\code{TRUE}).}
}
\description{
This function conducts a heuristic search for row subsets that minimize the
number of omitted rows that remove all missing values from a given matrix.
}

