% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paretoDat.R
\name{paretoDat}
\alias{paretoDat}
\title{Identify the Pareto maximum rows of a data frame (preserving ties)}
\usage{
paretoDat(dat, logic = FALSE)
}
\arguments{
\item{dat}{A data frame.}

\item{logic}{Logical; whether to return a logical vector of length the number
of entries in \code{dat} or a vector of entry indices of \code{dat}.}
}
\description{
Identify the Pareto maximum rows of a data frame (preserving ties)
}

